Add a testing script
I also found out the problem why I could not uses int constants as
"case" labels: It was because

const int somevar = somevalue;

is not a constant expression according to the C standard! Therfore
changed the constants into macros, and suddenly it compiled fine.

So far for my latest attempt to use nice clean constants instead of ugly
preprocessor macros: Not with C!