Fix C89 incompatibility
Replace printf format specifiers "%lc" by a replacement function which
converts the wide character into a MBCS string and then displays it as
a normal "%s".

I did not realize before that "%lc" is a C99 extension, and I definitely
do not want to rely on C99 features (yet): This standard is barely 17
years old, way to young for a conservative type like me! ;-)