Implement option arguments
Just now only the new -n option has any argument, but the code is
prepared to easily add more such options.

Rename the -x and -X options temporarily to -j and -J, and permanently
rename -m and -M into -x and -X.

Add option -a for turning -x and -X in what is -j and -J right now, but
will be dropped in favor of -a (which is not implemented yet) shortly.

The value accepted by -n is not used yet, either.

But update the help to explain what the new options should do
eventually.