Adapt -B and -X to new -n and -a
Everything seems to work now. Spaces characters (to be exact: unit
separators) between units in the numeric area are now completely
ignored, and so is everything in the ASCII dump, if any. There is also
no restriction any more about the number of units per line in the
numeric area - units are read until the end of the line or the separator
is reached (which separates the numeric area from the ASCII dump).

As a consequence, this tool can now also be misused as a pretty decent
kind of hexdump and un-hexdump utility, as well as a binary dump and
binary un-dump utility.

The only drawback compared to a real hexdump is that it does not display
any address offsets.