Add a generator for a large -w/-c test case
The generator creates all permutations of whitespace sequences which
might need encoding and a single word character which never needs
encoding.

The output of the generator is about 500 kB right now, and the testing
script has been extended with an "-e" option for using its outputs via
pipes only, avoiding the requirement to store the generated data in
full.

Actually, that new mode has been the result of a mistake: I originally
thought the generator would be 500 MB rather than kB, which would mean a
lot of wasted disk space.