Enhanced test case generator support
The -e option has been removed. In exchange for it, the -g option has
been added, which interprets all script arguments as generators.

All test will then be run, running all generators for each test.

In addition, a special mode for reading standard input has been added,
which only applies in the absence of option -g when no arguments are
specified. In this case, a single test case file is read from standard
input, and only a single test can be specified in this case.