Handle SIGINT separately
I still do not fully understand shell signal handling. A trap for EXIT
seems not to be taken in case of SIGINT.

But if the same signal handler is set for both SIGINT and SIGEXIT, it is
called twice in case of SIGINT!

A somewhat strange solution which seems to work is to set a separate
SIGINT handler which calls "exit". In this case, the SIGEXIT handler is
only called once in all cases.