Use enum rather than integers for states
Although it is rather pointless to think of good symbolic names for
states which are characterized by more semantics than a single name
could convey, it is easier to associate existing states with different
internal integer encodings when they have symbolic names.

This is especially useful of new states need to be added, but the new
state should be associated with an internal integer already in use.
Using enums, this can be achieved by just re-arranging the order of enum
symbols in the enum definition, without need to change anything else.

It is also easier to locate the usages of a state if it is associated
with a symbol rather than a small integer which might be found multiple
times in the source code without necessarily refering to a state.

Using enums as state labels also allows the GNU C compiler to check
whether all possible enum values have been handled in switch()
statements.

Also add an option -m to gen_evil, allowing to generate a much smaller
file for the basic potentially problematic cases, without generating all
possible permutations of them.

Finally, fix a bug that made -c start its output with a redundant
newline. This requires adding a new state.