Remove invalid optimization
It is not true that when tranforming 7 spaces before the end of a line,
this can result in a single literal space followed by 6 encoded spaces
and an encoded newline.

While it is correct that the encoded newline cannot be mis-interpreted
by that first literal space, the first of the encoded spaces would be
misinterpreted by the presence of the space before it, assuming to be
part of a different encoding.

Therefore, the optimization that only 6 spaces before an encoding need
to be encoded themselves has to be dropped.

All spaces prefixing an encoding need to be encoded themselves, too.

However, this seemed to be the last bug for now: The test cases
generated by gen_evil work fine now.